% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{clustering_energy}
\usage{
energy_clustering(data, clusters = 3, iter_max = 10, restarts = 1)
}
\arguments{
\item{data}{A gl object.}

\item{clusters}{Number of clusters.}

\item{iter_max}{Maximum number of iterations.}

\item{restarts}{Number of restarts.}
}
\value{
An object of class kgroups (see energy library).
}
\description{
Performs kgroups clustering by energy distance of biosensors.usc.
}
