% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_confidence_band}
\alias{cpp_confidence_band}
\title{This function computes intrinsic confidence bands for Wasserstein regression.}
\usage{
cpp_confidence_band(xfit, xpred, Q_obs, q_obs, t_vec, alpha)
}
\arguments{
\item{xfit}{A nxp matrix of predictor values for fitting (do not include a column for the intercept).}

\item{xpred}{A kxp matrix of input values for regressors for prediction.}

\item{Q_obs}{A nxm matrix of quantile functions. Q_obs(i, :) is a 1xm vector of quantile function values on grid t_vec.}

\item{q_obs}{A nxm matrix of quantile density functions. q_obs(i, :) is a 1xm vector of quantile density function values on grid t_vec.}

\item{t_vec}{A 1xm vector - common grid for all quantile density functions in Q_obs, q_obs, q_prime_obs.}

\item{alpha}{The significant level is 100*(1 - alpha).}
}
\value{
An object containing the components:
\code{Q_lx} Lower bound of confidence bands in terms of density functions.
\code{Q_ux} Upper bound of confidence bands in terms of density functions.
\code{Qpred} Fitted density function at xpred.
}
\description{
This function computes intrinsic confidence bands for Wasserstein regression.
}
