% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_wasserstein_regression}
\alias{cpp_wasserstein_regression}
\title{This function perform Frechet regression with the Wasserstein distance.}
\usage{
cpp_wasserstein_regression(xfit, q, Q0, xpred, t, qdmin)
}
\arguments{
\item{xfit}{A nxp matrix of predictor values for fitting (do not include a column for the intercept).}

\item{q}{A nxm matrix of quantile density functions. q(i, :) is a 1xm vector of quantile density function values on an equispaced grid on [0, 1].}

\item{Q0}{A 1xn array of quantile function values at 0.}

\item{xpred}{A kxp matrix of input values for regressors for prediction.}

\item{t}{A 1xm vector - common grid for all quantile density functions in q.  If missing, defaults to linspace(0, 1, m);  For best results, should use a finer grid than for quantle estimation, especially near the boundaries.}

\item{qdmin}{A positive lower bound on the estimated quantile densites. Defaults to 1e-6.}
}
\value{
An object containing the components:
\code{xpred} See input of same name.
\code{qpred} A kxN array.  qpred(l,:) is the regression prediction of q (the quantile density) given X = xpred(l, :)'
\code{Qpred} A kxm array. Qpred(l, :) is the regression prediction of Q given X = xpred(l, :)'
\code{fpred} A kxm array. fpred(l, :) is the regression prediction of f (the density) given X = xpred(l, :)', evaluated on the grid Qpred(l, :)
\code{xfit} See input of same name.
\code{qfit} A nxN array. qfit(l, :) is the regression prediction of q given X = xfit(l, :)'
\code{Qfit} A nxm array. Qfit(l, :) is the regression prediction of Q given X = xfit(l, :)'
\code{fpred} A kxm array. fpred(l, :) is the regression prediction of f (the density) given X = xpred(l, :)', evaluated on the grid Qfit(l, :)
\code{QP_used} A flag indicating whether OLS fits all satisfied the constraints (=0) or if the quadratic program was used in fitting (=1).
}
\description{
This function perform Frechet regression with the Wasserstein distance.
}
