% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{hypothesis_testing}
\alias{hypothesis_testing}
\title{hypothesis_testing}
\usage{
hypothesis_testing(data1, data2, permutations = 100)
}
\arguments{
\item{data1}{The glocodensities object of the first population.}

\item{data2}{The glocodensities object of the second population.}

\item{permutations}{The number of permutations.}
}
\value{
A biosensors.usc.comparison object with mean and variance of the two populations, and the energy and anova pvalues.
}
\description{
Loads the glocodensities from two csv files.
}
