% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(filename_fdata, filename_variables)
}
\arguments{
\item{filename_fdata}{The csv file with the functional data.}

\item{filename_variables}{The csv file with the clinical variables.}
}
\value{
A biosensors.usc object with the fdata densities and quantiles, and the dataframe with the variables.
}
\description{
Loads the glocodensities from two csv files.
}
\examples{
X = load_data("./test/data_1.csv", "./test/variables_1.csv")
}
