% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadayara.R
\name{nadayara_prediction}
\alias{nadayara_prediction}
\title{nadayara_prediction}
\usage{
nadayara_prediction(nadaraya, Qpred, hs = NULL)
}
\arguments{
\item{data}{A biosensor object.}

\item{predictor}{The name of the vector of observed values (one of the columns of data$variables).}
}
\value{
An object of class performance containing the components:
\code{call} The function call
\code{error} The residuals.
\code{prediction} The fitted regression.
\code{ub} The upper band of the regression.
\code{lb} The lower band of the regression.
where \code{error}, \code{prediction}, \code{ub}, and \code{lb} are \code{fdata} objects.
}
\description{
Performs an analysis of the performance of the regression vs a smoothing parameter.
}
