% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmod.R
\name{regmod_regression}
\alias{regmod_regression}
\title{regmod regression}
\usage{
regmod_regression(data, predictor)
}
\arguments{
\item{data}{A gd object.}

\item{response}{The name of the vector of observed values (one of the columns of data$variables).}
}
\value{
An object of class wasserstein containing the components:
\code{call} The function call.
\code{error} The residuals.
\code{prediction} The fitted regression.
\code{ub} The upper band of the regression.
\code{lb} The lower band of the regression.
where \code{error}, \code{prediction}, \code{ub}, and \code{lb} are \code{fdata} objects.
}
\description{
Performs the Wasserstein regression.
}
