% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge_prediction}
\alias{ridge_prediction}
\title{ridge_prediction}
\usage{
ridge_prediction(m, X, Xpred, method = "manhattan", type = "gaussian")
}
\arguments{
\item{m}{ToDo.}

\item{X}{ToDo.}

\item{Xpred}{ToDo.}

\item{method}{The distance measure to be used (@seealso parallelDist::parDist)}

\item{type}{The kernel type ("gaussian" or "lapla")}
}
\value{
An object of class wasserstein containing the components:
\code{call} The function call.
\code{error} The residuals.
\code{prediction} The fitted regression.
\code{ub} The upper band of the regression.
\code{lb} The lower band of the regression.
}
\description{
Generates a quantile reression model V + V2 * v + tau * V2 * Q0 where Q0 is a truncated random variable, v = 2 * X, tau = 2 * X, V ~ Unif(-1, 1), V2 ~ Unif(-1, -1), V3 ~ Unif(0.8, 1.2), and E(V|X) = tau * Q0;
}
