% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein.R
\name{wasserstein_regression}
\alias{wasserstein_regression}
\title{wasserstein_regression}
\usage{
wasserstein_regression(data, response)
}
\arguments{
\item{data}{A biosensor object.}

\item{response}{The name of the scalar response. The response must be a column name in data$variables.}
}
\value{
An object of class wasserstein containing the components:
\code{prediction} The fitted regression.
\code{regression} An internal bwasserstein object (@seealso cpp_wasserstein_regression)
\code{data} A data frame with biosensor raw data.
\code{response} The name of the scalar response.
}
\description{
Performs the Wasserstein regression using a quantile density function.
}
\examples{
# Data extracted from the paper: Hall, H., Perelman, D., Breschi, A., Limcaoco, P., Kellogg, R.,
# McLaughlin, T., Snyder, M., Glucotypes reveal new patterns of glucose dysregulation, PLoS
# biology 16(7), 2018.
file1 = system.file("extdata", "data_1.csv", package = "biosensors.usc")
file2 = system.file("extdata", "variables_1.csv", package = "biosensors.usc")
data = load_data(file1, file2)
wass = wasserstein_regression(data, "BMI")
}
