% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_prediction}
\alias{clustering_prediction}
\title{clustering_prediction}
\usage{
clustering_prediction(clustering, objects)
}
\arguments{
\item{clustering}{A gl.clustering object.}

\item{objects}{Matrix of objects to cluster.}
}
\value{
The clusters to which these objects are assigned.
}
\description{
Predicts the cluster of each element of the objects list
}
